+++ src/operator/tensor/indexing_op.h
@@ -670,209 +670,6 @@ void OneHotOpForward(const nnvm::NodeAttrs& attrs,
   });
 }
 
-/*!
- * \brief sparse retain namespace
- */
-namespace sr {
-enum SparseRetainOpInputs {kArr, kIdx};
-enum SparseRetainOpOutputs {kOut};
-}  // namespace sr
-
-inline bool SparseRetainOpShape(const nnvm::NodeAttrs& attrs,
-                                std::vector<TShape> *in_attrs,
-                                std::vector<TShape> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), 2U)
-    << "sparse_retain operator takes 2 arguments (" << in_attrs->size() << " given)";
-  CHECK_EQ(out_attrs->size(), 1U);
-
-  TShape tshape((*in_attrs)[sr::kArr]);
-  shape_assign(&tshape, (*out_attrs)[sr::kOut]);
-  SHAPE_ASSIGN_CHECK(*in_attrs, sr::kArr, tshape);
-  SHAPE_ASSIGN_CHECK(*out_attrs, sr::kOut, tshape);
-  return true;
-}
-
-inline bool SparseRetainOpType(const nnvm::NodeAttrs& attrs,
-                               std::vector<int> *in_attrs,
-                               std::vector<int> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), 2U);
-  CHECK_EQ(out_attrs->size(), 1U);
-  CHECK_NE((*in_attrs)[sr::kIdx], -1) << "Index type must be set for sparse_retain operator";
-
-  TYPE_ASSIGN_CHECK(*out_attrs, 0, (*in_attrs)[sr::kArr]);
-  TYPE_ASSIGN_CHECK(*in_attrs, 0, (*out_attrs)[sr::kOut]);
-  return (*in_attrs)[0] != -1;
-}
-
-inline bool SparseRetainForwardInferStorageType(const nnvm::NodeAttrs& attrs,
-                                                const Context& ctx,
-                                                std::vector<int> *in_attrs,
-                                                std::vector<int> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), 2U);
-  CHECK_EQ(out_attrs->size(), 1U);
-  STORAGE_TYPE_ASSIGN_CHECK(*in_attrs, sr::kArr, kRowSparseStorage);
-  STORAGE_TYPE_ASSIGN_CHECK(*out_attrs, sr::kOut, kRowSparseStorage);
-  return true;
-}
-
-inline bool SparseRetainBackwardInferStorageType(const nnvm::NodeAttrs& attrs,
-                                                 const Context& ctx,
-                                                 std::vector<int> *in_attrs,
-                                                 std::vector<int> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), 2U);
-  CHECK_EQ(out_attrs->size(), 2U);
-  STORAGE_TYPE_ASSIGN_CHECK(*in_attrs, sr::kOut, kDefaultStorage);
-  STORAGE_TYPE_ASSIGN_CHECK(*in_attrs, sr::kIdx, kDefaultStorage);
-  STORAGE_TYPE_ASSIGN_CHECK(*out_attrs, sr::kArr, kRowSparseStorage);
-  STORAGE_TYPE_ASSIGN_CHECK(*out_attrs, sr::kIdx, kDefaultStorage);
-  return true;
-}
-
-struct SparseRetainRspForward {
-  template<typename DType, typename RType, typename IType>
-  MSHADOW_XINLINE static void Map(int i, DType* out_data, RType* out_idx,
-                                  const DType* in_data, const RType* in_idx,
-                                  const IType* idx, const size_t nnr,
-                                  const size_t num_cols) {
-    const RType irow = idx[i];
-    int j = -1, left = 0, right = nnr - 1;
-    while (left <= right) {
-      int m = left + (right - left) / 2;
-      const auto in_idx_m = in_idx[m];
-      if (in_idx_m == irow) {
-        j = m;
-        break;
-      } else if (in_idx_m < irow) {
-        left = m + 1;
-      } else {
-        right = m - 1;
-      }
-    }
-    out_idx[i] = idx[i];
-    if (j >= 0) {
-      const size_t in_offset = j * num_cols;
-      const size_t out_offset = i * num_cols;
-      for (size_t k = 0; k < num_cols; ++k) {
-        out_data[out_offset+k] = in_data[in_offset+k];
-      }
-    }
-  }
-};
-
-template<typename xpu>
-void SparseRetainOpForwardRspImpl(mshadow::Stream<xpu> *s, const NDArray &input,
-                                  const TBlob &indices, OpReqType req,
-                                  NDArray *output) {
-  using namespace rowsparse;
-  if (req == kNullOp || !input.storage_initialized() || indices.Size() == 0U) {
-    FillZerosRspImpl(s, output);
-    return;
-  }
-  const TBlob input_data = input.data();
-  const TBlob input_idx = input.aux_data(rowsparse::kIdx);
-  output->CheckAndAlloc({mshadow::Shape1(indices.Size())});
-  TBlob output_data = output->data();
-  TBlob output_idx = output->aux_data(rowsparse::kIdx);
-  const index_t row_length = input_data.shape_.ProdShape(1, input_data.shape_.ndim());
-
-  using namespace mxnet_op;
-  MSHADOW_TYPE_SWITCH(output_data.type_flag_, DType, {  // output data type
-    MSHADOW_IDX_TYPE_SWITCH(output_idx.type_flag_, RType, {  // row index data type
-      MSHADOW_TYPE_SWITCH(indices.type_flag_, IType, {  // index array data type
-        Kernel<set_zero, xpu>::Launch(s, output_data.Size(), output_data.dptr<DType>());
-        Kernel<SparseRetainRspForward, xpu>::Launch(s, indices.Size(), output_data.dptr<DType>(),
-            output_idx.dptr<RType>(), input_data.dptr<DType>(), input_idx.dptr<RType>(),
-            indices.dptr<IType>(), input_data.shape_[0], row_length);
-      });
-    });
-  });
-}
-
-template<typename xpu>
-void SparseRetainOpForwardEx(const nnvm::NodeAttrs& attrs,
-                             const OpContext& ctx,
-                             const std::vector<NDArray>& inputs,
-                             const std::vector<OpReqType>& req,
-                             const std::vector<NDArray>& outputs) {
-  CHECK_EQ(inputs.size(), 2U);
-  CHECK_EQ(outputs.size(), 1U);
-  CHECK_EQ(req.size(), 1U);
-  CHECK_EQ(req[sr::kOut], kWriteTo) << "sparse_retain only supports req=\'write\'";
-
-  CHECK_EQ(inputs[sr::kArr].storage_type(), kRowSparseStorage)
-    << "sparse_retain operator only takes row sparse NDArray as input";
-  CHECK_EQ(inputs[sr::kIdx].storage_type(), kDefaultStorage)
-    << "sparse_retain operator only takes default NDArray as its index array";
-  CHECK_EQ(outputs[sr::kOut].storage_type(), kRowSparseStorage)
-    << "sparse_retain operator only outputs row sparse NDArray";
-
-  const NDArray& input_nd = inputs[sr::kArr];
-  const TBlob idx_data = inputs[sr::kIdx].data();
-  NDArray output_nd = outputs[sr::kOut];
-  mshadow::Stream<xpu> *s = ctx.get_stream<xpu>();
-  SparseRetainOpForwardRspImpl<xpu>(s, input_nd, idx_data, req[sr::kOut], &output_nd);
-}
-
-template<int req>
-struct SparseRetainRspBackward {
-  template<typename DType, typename RType, typename IType>
-  MSHADOW_XINLINE static void Map(int i, DType* in_grad, RType* in_grad_idx,
-                                  const DType* out_grad, const IType* idx,
-                                  const size_t num_cols) {
-    const RType irow = idx[i];
-    in_grad_idx[i] = irow;
-    const size_t out_offset = irow * num_cols;
-    const size_t in_offset = i * num_cols;
-    for (size_t j = 0; j < num_cols; ++j) {
-      KERNEL_ASSIGN(in_grad[in_offset+j], req, out_grad[out_offset+j]);
-    }
-  }
-};
-
-template<typename xpu>
-void SparseRetainOpBackwardEx(const nnvm::NodeAttrs& attrs,
-                              const OpContext& ctx,
-                              const std::vector<NDArray>& inputs,
-                              const std::vector<OpReqType>& req,
-                              const std::vector<NDArray>& outputs) {
-  CHECK_EQ(inputs.size(), 2U);
-  CHECK_EQ(outputs.size(), 2U);
-  CHECK_EQ(req.size(), 2U);
-  CHECK_NE(req[sr::kArr], kWriteInplace);
-  CHECK_EQ(req[sr::kIdx], kNullOp)
-    << "sparse_retain does not support calculating gradients of indices";
-
-  CHECK_EQ(inputs[sr::kOut].storage_type(), kDefaultStorage)
-    << "sparse_retain backward only takes default NDArray as ograd";
-  CHECK_EQ(inputs[sr::kIdx].storage_type(), kDefaultStorage)
-    << "sparse_retain backward only takes default NDArray as its index array";
-  CHECK_EQ(outputs[sr::kArr].storage_type(), kRowSparseStorage)
-    << "sparse_retain backward only outputs row sparse NDArray as grad of input";
-
-  const TBlob out_grad_data = inputs[sr::kOut].data();
-  const TBlob idx_data = inputs[sr::kIdx].data();
-
-  NDArray in_grad_nd = outputs[sr::kArr];
-  in_grad_nd.CheckAndAlloc({mshadow::Shape1(idx_data.Size())});
-  TBlob in_grad_data = in_grad_nd.data();
-  TBlob in_grad_idx = in_grad_nd.aux_data(rowsparse::kIdx);
-  const auto row_length = out_grad_data.shape_.ProdShape(1, out_grad_data.shape_.ndim());
-
-  using namespace mxnet_op;
-  Stream<xpu> *s = ctx.get_stream<xpu>();
-  MSHADOW_TYPE_SWITCH(out_grad_data.type_flag_, DType, {  // output data type
-    MSHADOW_IDX_TYPE_SWITCH(in_grad_idx.type_flag_, RType, {  // row index data type
-      MSHADOW_TYPE_SWITCH(idx_data.type_flag_, IType, {  // index array data type
-        MXNET_ASSIGN_REQ_SWITCH(req[sr::kArr], req_type, {
-          Kernel<SparseRetainRspBackward<req_type>, xpu>::Launch(
-              s, in_grad_idx.Size(), in_grad_data.dptr<DType>(), in_grad_idx.dptr<RType>(),
-              out_grad_data.dptr<DType>(), idx_data.dptr<IType>(), row_length);
-        });
-      });
-    });
-  });
-}
-
 }  // namespace op
 }  // namespace mxnet
 #ifdef __CUDACC__
