+++ tests/python/unittest/test_sparse_operator.py
@@ -171,9 +171,9 @@ def test_sparse_slice():
 
 
 def test_sparse_retain():
-    def check_sparse_retain(shape):
+    def check_sparse_retain(shape, density):
         num_rows = shape[0]
-        rsp, _ = rand_sparse_ndarray(shape=shape, stype='row_sparse', density=0.5)
+        rsp, _ = rand_sparse_ndarray(shape=shape, stype='row_sparse', density=density)
         length = np.random.randint(1, num_rows + 1)
         idx = random_sample(list(range(0, num_rows)), length)
         idx.sort()
@@ -189,11 +189,15 @@ def test_sparse_retain():
         data = mx.symbol.Variable('data')
         idx = mx.symbol.Variable('indices')
         sym = mx.sym.sparse_retain(data=data, indices=idx)
-        check_numeric_gradient(sym, [rsp, indices], grad_nodes=['data'], grad_stype_dict={'data': 'row_sparse'})
+        check_numeric_gradient(sym, [rsp, indices], grad_nodes=['data'],
+                               grad_stype_dict={'data': 'row_sparse'})
+
     shape = rand_shape_2d()
     shape_3d = rand_shape_3d()
-    check_sparse_retain(shape)
-    check_sparse_retain(shape_3d)
+    densities = [0.01, 0.1, 0.2, 0.5, 0.8, 1.0]
+    for density in densities:
+        check_sparse_retain(shape, density)
+        check_sparse_retain(shape_3d, density)
 
 
 def test_sparse_nd_zeros():
